.github: Standup a barebones MySQL database instance.
These are fairly-straightfoward changes for standing up a mysql database that contains the timbot schema.sql and ensures that schema gets loaded correctly by the default ubuntu mysql instance (5.7). Was having difficulties getting this to work with any mysql containers and there's no great way to stop the timbot container while running the github jobs, so this will have to due for now.